Game 6 : Menu - Added Menu UI - Added Menu animations
Game 6 : Upgrades
- Added Upgrade Menu UI
- Added an upgrade system where a turret can be selected and upgraded
- Added Upgrade Menu Ui animations

+++ VolaTileU/Assets/Scripts/ScriptsGame6/Game6EnemyHealth.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class Game6EnemyHealth : MonoBehaviour
 {
     [Header("Attributes")]
-    [SerializeField] private int hitPoints = 2;
+    [SerializeField] private float hitPoints = 2;
     [SerializeField] private int currencyWorth = 50;
 
     private bool isDestroyed = false;
 
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         hitPoints -= dmg;
